requirements.txt
flask==2.2.3
gunicorn==20.1.0
lark-oapi==1.0.21
requests==2.28.2
python-dotenv==1.0.0
pillow==9.4.0
numpy==1.24.2
pandas==1.5.3
pytest==7.3.1
supervisor==4.2.5
flask-cors==3.0.10
pycryptodome==3.17.0
redis==4.5.1 